gptengine-app/requirements.txt
Flask
Flask-Cors
Flask-RESTful
flask-restplus
flask-redis
flasgger
environs
requests
Flask-Migrate
marshmallow_sqlalchemy
marshmallow-sqlalchemy
flask-marshmallow
celery
openpyxl
pyquery
pymysql
captcha
pandas
